requirements.txt
pyTelegramBotAPI==4.12.0
python-dotenv==0.21.1
psycopg2-binary==2.9.9
